+++ src/validators/index.ts
@@ -1,10 +1,9 @@
-export * from './registerDataValidator';
 export * from './filterParametresValidator';
-export * from './insertedQuestionValidator';
 export * from './emailValidator';
-export * from './lessonValidator';
-export * from './updateDataValidator';
+export * from './lesson/lessonValidator';
 export * from './lesson';
 export * from './course';
 export * from './module';
 export * from './group';
+export * from './question';
+export * from './user';
