+++ src/validators/lesson/index.ts
@@ -2,3 +2,4 @@ export * from './lessonFilterParametresValidator';
 export * from './lessonUpdateDataValidator';
 export * from './addQuestionToLessonValidator';
 export * from './lessonPassedTestDataValidator';
+export * from './lessonValidator';
