+++ src/validators/question/index.ts
@@ -1,0 +1,2 @@
+export * from './insertedQuestionValidator';
+export * from './questionIdValidator';
