+++ src/validators/question/questionIdValidator.ts
@@ -1,0 +1,5 @@
+import * as Joi from 'joi';
+
+export const questionIdValidator = Joi.object().keys({
+    question_id: Joi.number().integer().required()
+});
