+++ src/validators/user/index.ts
@@ -1,0 +1,3 @@
+export * from './userIdValidator';
+export * from './registerDataValidator';
+export * from './updateDataValidator';
