+++ src/validators/user/updateDataValidator.ts
@@ -1,6 +1,6 @@
 import * as Joi from 'joi';
 
-import { RegExpEnum } from '../constants';
+import { RegExpEnum } from '../../constants';
 
 export const updateDataValidator = Joi.object().keys({
   email: Joi.string().max(255).regex(RegExpEnum.email).trim(),
