+++ src/validators/user/userIdValidator.ts
@@ -1,0 +1,5 @@
+import * as Joi from 'joi';
+
+export const userIdValidator = Joi.object().keys({
+    user_id: Joi.number().integer().required()
+});
